requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.5
django-localflavor==3.1
django-recaptcha==2.0.6
django-sort-order-field==1.2
gunicorn==20.1.0
psycopg2==2.9.1
python-decouple==3.4
python-stdnum==1.16
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
whitenoise==5.3.0
